+++ src/components/button/button.ts
@@ -116,13 +116,8 @@ export default class WaButton extends WebAwesomeFormAssociatedElement {
   /** Tells the browser where to open the link. Only used when `href` is present. */
   @property() target: '_blank' | '_parent' | '_self' | '_top';
 
-  /**
-   * When using `href`, this attribute will map to the underlying link's `rel` attribute. Unlike regular links, the
-   * default is `noreferrer noopener` to prevent security exploits. However, if you're using `target` to point to a
-   * specific tab/window, this will prevent that from working correctly. You can remove or change the default value by
-   * setting the attribute to an empty string or a value of your choice, respectively.
-   */
-  @property() rel = 'noreferrer noopener';
+  /** When using `href`, this attribute will map to the underlying link's `rel` attribute. */
+  @property() rel?: string;
 
   /** Tells the browser to download the linked file as this filename. Only used when `href` is present. */
   @property() download?: string;
@@ -281,7 +276,7 @@ export default class WaButton extends WebAwesomeFormAssociatedElement {
         href=${ifDefined(isLink ? this.href : undefined)}
         target=${ifDefined(isLink ? this.target : undefined)}
         download=${ifDefined(isLink ? this.download : undefined)}
-        rel=${ifDefined(isLink ? this.rel : undefined)}
+        rel=${ifDefined(isLink && this.rel ? this.rel : undefined)}
         role=${ifDefined(isLink ? undefined : 'button')}
         aria-disabled=${this.disabled ? 'true' : 'false'}
         tabindex=${this.disabled ? '-1' : '0'}