+++ src/components/dialog/dialog.ts
@@ -95,7 +95,7 @@ export default class WaDialog extends WebAwesomeElement {
   disconnectedCallback() {
     super.disconnectedCallback();
     unlockBodyScrolling(this);
-    this.closeWatcher?.destroy();
+    this.removeOpenListeners();
   }
 
   private async requestClose(source: Element) {